UUFV-380 Refactor PeriodsSectionImage
