+++ libs/application/templates/parental-leave/src/fields/ConclusionSectionImage/ConclusionImageScreen.css.ts
