--- libs/application/templates/parental-leave/src/fields/PeriodsSectionImage/ManWithStrollerIllustration.tsx
@@ -1,100 +1,0 @@
-import React from 'react'
-
-// TODO later move the illustrations from the web project and into a reusable library
-function ManWithStrollerIllustration() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 546 563">
-      <defs>
-        <linearGradient
-          id="linear-gradient"
-          x1="327.312"
-          x2="179.115"
-          y1="457.283"
-          y2="200.598"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0" stopColor="#ff4c84" />
-          <stop offset="0.136" stopColor="#ee5390" />
-          <stop offset="0.406" stopColor="#c164af" />
-          <stop offset="0.782" stopColor="#797fe0" />
-          <stop offset="1" stopColor="#4c90ff" />
-        </linearGradient>
-        <style>
-          {`.cls-3{fill:none;stroke:#6a2ea0;stroke-linecap:round;stroke-linejoin:round;stroke-width:3px}.cls-4{fill:#0061ff}.cls-5{fill:#6a2ea0}`}
-        </style>
-      </defs>
-      <path
-        fill="url(#linear-gradient)"
-        d="M217.808 298.879l30.384-.025-22 38.873h-99.427s-22.4-22.68-20.1-59.885 38.4-67.85 75.122-67.32v52.367a35.991 35.991 0 0036.021 35.99zm181.866 116.626l.912-1.192L409.5 403.5l-33.556-31.188s3.715-52.651-4.609-69.687c0 0-17.268 1.311-26.212 4.522s-17.586.72-19.035-.722l-36.113 99.048 15.393 7.642 34.9-63.692s12.183 30.879 15.17 33.633 44.5 31.916 44.5 31.916-.115.207-.264.533z"
-      />
-      <path
-        fill="#ff4c84"
-        d="M265.4 412.2a23.347 23.347 0 11-23.347-23.347A23.373 23.373 0 01265.4 412.2z"
-      />
-      <path
-        d="M457.741 277.797v84.71m0-120.807s-32.778 33.3 0 75.271c32.167-43.412 0-75.271 0-75.271z"
-        className="cls-3"
-      />
-      <path
-        d="M266.945 71.382l26.185-26.176a19.171 19.171 0 10-26.185 26.176z"
-        className="cls-4"
-      />
-      <path
-        d="M319.571 18.872l-82.186 81.543m66.389-63.501A32.194 32.194 0 01258.9 80.837"
-        className="cls-3"
-      />
-      <circle cx="215.385" cy="122.243" r="4.604" className="cls-4" />
-      <circle cx="457.741" cy="218.243" r="4.604" className="cls-4" />
-      <circle cx="263.385" cy="506.243" r="4.604" className="cls-4" />
-      <circle cx="297.261" cy="275.021" r="6.149" className="cls-5" />
-      <path d="M45.385 362.507h74.338m256.221 0h123.45" className="cls-3" />
-      <circle cx="297.03" cy="275.357" r="6.169" className="cls-5" />
-      <circle cx="124.129" cy="419.912" r="2.301" className="cls-4" />
-      <circle cx="297.27" cy="274.282" r="5.907" className="cls-5" />
-      <circle cx="129.561" cy="413.392" r="12.138" className="cls-4" />
-      <circle cx="129.561" cy="413.392" r="22.247" className="cls-3" />
-      <path
-        d="M98.967 299.671h161.241l-21.568 39.003H108.874"
-        className="cls-3"
-      />
-      <path
-        d="M108.874 338.675c-14.011-14.969-21.556-35.086-21.556-57.208 0-46.244 38.5-89.264 95.022-89.264m-.014 33.761l.008 37.606c0 20.516 16.644 36.129 37.161 36.113l32.459-.025m-46.529 39.494l-75.501 74.24"
-        className="cls-3"
-      />
-      <path
-        d="M294.627 274.282h-13.213a35.993 35.993 0 00-30.495 16.871l-5.341 8.518m-95.988 39.004l92.459 74.673M133.8 252.214s19.01-29.856 64.546-28.035m126.742 82.246l-35.954 98.7c-5.91-2.92-8.608 5.162-4.928 7.5 3.958 2.516 16.535 8.934 16.535 8.934l39.525-72.137"
-        className="cls-3"
-      />
-      <path
-        d="M296.374 419.526l-6.884 14.032-22.638-9.541s4.312-8.855 17.43-2.68l3.2-6.8m-82.057 19.976h99.68m80.196 0h39.253m-11.262-24.029l13.616 10.916-12.844 13.115h-11.526s-.968-6.4 6.257-6.524l3.082-5.42-4.788-4.266"
-        className="cls-3"
-      />
-      <path
-        d="M318.891 301.616s8.594 11.454 26.289 5.384a76.917 76.917 0 0126.155-4.374s8.783 20.972 3.719 68.86l39.718 36.914s-6.529 8.624-8.951 11.478c-3.283 3.87-10.364-.337-5.888-4.905 0 0-40.413-26.62-43.4-31.069-9.42-14.61-23.3-52.68-23.3-52.68m38.102-28.599s-4.38-19.494.337-48.785c3.183-19.77-9.384-32.749-14.391-37.047a4.8 4.8 0 01-1.643-4.28 31.62 31.62 0 011.286-6.37m-25.642 59.216a57.614 57.614 0 00.32-11.855c-1.135-12.434 1.426-20.789 5.811-25.845a18.452 18.452 0 011.8-1.8l-3.826-8.135m-11.928 77.931a149.286 149.286 0 005.027-15.874"
-        className="cls-3"
-      />
-      <path
-        d="M364.071 251.343c-1.383 5.974-4.736 16.557-12.25 24.869-24.561 6.563-39.687 1.037-49.659-3.129"
-        className="cls-3"
-      />
-      <path
-        d="M305.5 266.867s3.8-.54 9.216-.974c8.161-.653 19.991-1.065 28.018.974m1.632-19.206a153.474 153.474 0 00-4.1 18.233m-8.127-83.044s-9.137 10.914-5.513 25.484"
-        className="cls-3"
-      />
-      <path
-        d="M324.847 200.412s-3.83-.382-3.9-3.334c-.05-2.174 7.6-8.186 7.6-8.186m-39.529 77.975s7.065-9.207 16.485 0"
-        className="cls-3"
-      />
-      <path
-        d="M331.282 170.846a12.916 12.916 0 0011.856 5.073c8.03-.767 22.311.279 22.343 14.407s-7.981 24.827-7.981 24.827-9.163-.827-10.416-6.294c0 0 5.394-8.3.578-9.24s-8.146 6.768-7.4 17.486c0 0-3.625 4.36-11.714 2.752s-3.989-14.349-3.989-14.349 5.613 2.677 10.822.037c2-1.809 5.65-12.454 5.57-15.219-8.812 0-19.263-8.178-9.669-19.48z"
-        className="cls-4"
-      />
-      <path
-        d="M335.824 230.765s14.427 4.229 23.676-9.908m-44.781 46.036v10.488"
-        className="cls-3"
-      />
-    </svg>
-  )
-}
-
-export default ManWithStrollerIllustration
