--- libs/application/templates/parental-leave/src/fields/PeriodsSectionImage/WomanWithLaptopIllustration.tsx
@@ -1,68 +1,0 @@
-import React from 'react'
-
-function WomanWithLaptopIllustration() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 546 563">
-      <defs>
-        <linearGradient
-          id="linear-gradient"
-          x1="49.222"
-          x2="381.811"
-          y1="211.727"
-          y2="544.316"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0" stopColor="#ff4c84"></stop>
-          <stop offset="0.354" stopColor="#b568b7"></stop>
-          <stop offset="0.657" stopColor="#7c7ede"></stop>
-          <stop offset="0.88" stopColor="#598bf6"></stop>
-          <stop offset="1" stopColor="#4c90ff"></stop>
-        </linearGradient>
-        <style>
-          {`.cls-1{fill:none;stroke:#6a2ea0;stroke-linecap:round;stroke-linejoin:round;stroke-width:3px}.cls-4{fill:#0061ff}`}
-        </style>
-      </defs>
-      <circle cx="127.72" cy="157.13" r="23.952" className="cls-1"></circle>
-      <path
-        fill="url(#linear-gradient)"
-        d="M334.948 383.224a12.076 12.076 0 11-12.076-12.076 12.076 12.076 0 0112.076 12.076zm-170.841 51.018H134.58c-1.757 0-36.673-1.261-36.673-1.261s34.7-6.921 38.5-8.365c21.537-4.039 57.881-31.3 57.881-31.3 23.89 11.432 38.363 2.152 38.363 2.152a22.839 22.839 0 008.091 5.314l5.963 21.187h-70.33a12.271 12.271 0 00-12.268 12.273zm-80.194-9.627a26.249 26.249 0 0126.249-26.248c.654 0 1.3.032 1.943.079l80.84-22.962 29.95 2.355 15.647 15.12-23.217-82.492a36.89 36.89 0 00-35.51-26.895h-4.2v28.464a14.639 14.639 0 01-14.639 14.639 14.639 14.639 0 01-14.639-14.639v-28.464h-1.88a38.466 38.466 0 00-36.16 25.345L72.919 406.4a20.778 20.778 0 0012.364 26.583 26.2 26.2 0 01-1.37-8.368zm370.927 19.606v19.021c0 8.663-10.966 19-10.966 19h-24.44s-10.966-10.338-10.966-19v-19.021h46.372zM220.2 232.1c-13.546-1.879-30.574-24.277-32.656-37.971.287 29.377-39.724 37.971-39.724 37.971s-.044-5.8.029-9.2c.014-.622-12.873-2.39-14.36 6.327-1.066 6.248 4.09 17.544 12.851 17.182l.652 8.715a47.446 47.446 0 01-27.037-75.118 23.936 23.936 0 1138.073-19.31c0 .237-.011.472-.018.708 52.735-1.424 69.208 48.684 62.19 70.696z"
-      ></path>
-      <path
-        fill="#ff4c84"
-        d="M295.1 98.592a47.416 47.416 0 0070.322 62.135z"
-      ></path>
-      <circle cx="407.5" cy="242.243" r="4.604" className="cls-4"></circle>
-      <circle cx="359.5" cy="242.243" r="4.604" className="cls-4"></circle>
-      <circle cx="288.981" cy="242.243" r="4.604" className="cls-4"></circle>
-      <path
-        d="M175.614 283.648s30.757-2.648 38.346 23.306 23.74 84.262 23.74 84.262"
-        className="cls-1"
-      ></path>
-      <path
-        d="M146.337 293.908v6.7a14.639 14.639 0 0014.639 14.639 14.638 14.638 0 0014.638-14.639v-28.039m-77.707 129.5l95.257-26.895 30.105 2.665 19.481 18.529m-78.643-46.839a14.764 14.764 0 01-14.764 14.764 14.764 14.764 0 01-14.763-14.764m75.258-.567a14.745 14.745 0 01-6.359 12.139l2.417 14.84m-86.609-7.861l-9.125 30.287m-59.671 35.876h181.636m149.457 0H232.216l28.149-96.849h149.368l-28.149 96.849zm-168.079-20.462l3.333 20.461M365.323 151.07a40.764 40.764 0 1110.077-26.827m31.04 327.242h25.214m-56.259 30.758h101.9m-283.91 0h94"
-        className="cls-1"
-      ></path>
-      <circle cx="459.896" cy="460.243" r="8.757" className="cls-1"></circle>
-      <path
-        d="M333.385 74.243v50l18.659 16.075m6.869 120.998h-20.128a19.113 19.113 0 01-19.113-19.116 19.113 19.113 0 0119.113-19.113h93.756a19.113 19.113 0 0119.113 19.113 19.113 19.113 0 01-19.113 19.113h-49.32m-15.383 52.93l-9.881 33.994m25.264-86.921l-8.408 28.927m30.687 0h3a25 25 0 0125 25V334.3a25.939 25.939 0 0025.939 25.939H473.7a14 14 0 0114 14 14 14 0 01-14 14h-21.08a21.124 21.124 0 00-21.12 21.127v22.877m-96.115-142h-47.996M145.878 255.6q-1.369.09-2.761.09a41.836 41.836 0 110-83.672m60.401 63.289a78.927 78.927 0 01-2.5 11.467c-8.416 28.286-35.335 25.186-35.335 25.186"
-        className="cls-1"
-      ></path>
-      <path
-        d="M146.337 294.032v6.7a14.639 14.639 0 0014.639 14.638 14.638 14.638 0 0014.638-14.638v-28.04m12.221-49.183s1.55 10.036 1.79 15.363a4.261 4.261 0 01-5.86 4.34 7.827 7.827 0 01-3.056-2.379m-12.573 11.021s2.26 7.376 9.832 7.664"
-        className="cls-1"
-      ></path>
-      <path
-        d="M146.337 246.774v37s-23.219-2.221-34.325 17.97c-7.851 14.276-27.816 72.965-38.839 106.126a18.643 18.643 0 0018.492 24.5c24.636-1.006 63.539-8.374 102.626-39.232 0 0 23.725 11.513 36.849 2.463"
-        className="cls-1"
-      ></path>
-      <circle cx="167.911" cy="223.509" r="11.798" className="cls-1"></circle>
-      <circle cx="205.44" cy="223.509" r="11.798" className="cls-1"></circle>
-      <path
-        d="M179.709 223.509h13.933m-37.528-1.418h-16.478a6.872 6.872 0 00-6.872 6.872"
-        className="cls-1"
-      ></path>
-    </svg>
-  )
-}
-
-export default WomanWithLaptopIllustration
