+++ libs/application/templates/parental-leave/src/fields/index.ts
@@ -16,9 +16,8 @@ export { default as Rights } from './Rights/Rights'
 export { Review } from './Review/Review'
 export { default as EditOrAddEmployersAndPeriodsReview } from './EditOrAddEmployersAndPeriodsReview'
 export { default as InReviewSteps } from './InReviewSteps'
-export { default as Conclusion } from './PeriodsSectionImage/ConclusionImageScreen'
+export { default as Conclusion } from './ConclusionSectionImage/ConclusionImageScreen'
 export { default as DraftRequiresAction } from './DraftRequiresAction'
-export { default as PeriodsSectionImage } from './PeriodsSectionImage/PeriodsSectionImage'
 export { default as ExistingApplications } from './ExistingApplications'
 export { default as EditsInReviewSteps } from './InReviewSteps/EditsInReviewSteps'
 export { default as EditsRequireAction } from './EditsRequireAction'
